Refactored specific mission chart into own component
